--- sunkit_instruments/abstractions/__init__.py
@@ -1,6 +1,0 @@
-"""
-A subpackage for abstract classes for defining instrument-specific
-properties.
-"""
-
-from sunkit_instruments.abstractions.channels import *  # pragma: noqa
