+++ sunkit_instruments/response/__init__.py
@@ -1,0 +1,5 @@
+"""
+A subpackage for computing instrument responses
+"""
+
+from sunkit_instruments.response.thermal import *
